# Improve evalscript performance