--- a/tests/integration_tests/test_integration.py
+++ b/tests/integration_tests/test_integration.py
@@ -32,6 +32,8 @@ from tests.utils import (
         ),
         ("test_count_with_condition", [{"B01": 0, "B02":1}, {"B01":2, "B02":3}, {"B01":4, "B02":5}, {"B01":None, "B02": None}], [1, 2]),
         ("test_count_without_condition", [{"B01": 0, "B02":1}, {"B01":2, "B02":3}, {"B01":4, "B02":5}, {"B01":None, "B02": 3}], [3, 4]),
+        ("test_graph_1", [], [None]),
+        ("test_mean_of_mean", [{"B04": 0, "B08":1}, {"B04":2, "B08":3}, {"B04":3, "B08":5}, {"B04":1, "B08": 4}], [2.375]),
     ],
 )
 def test_convertable_process_graphs(pg_name, example_input, expected_output):