@Ardweaden 2022-03-18
> Comments are again purely cosmetical so fftr. You probably also need a computer upgrade as they are ( also on the master) much faster on mine

Haha it has indeed been struggling recently, sometime running eobrowser locally slowed it down to a crawl 😆.

Anyway, the difference can also be checked in EOB. Here are the evalscript generated from `reduce_mean_one_band`. I have only changed the `sampleType` to `AUTO` and set output bands to 1 in `updateOutput`. I set the timespan to 1st to 14th march.

Old: https://sentinelshare.page.link/L5oz
New: https://sentinelshare.page.link/62xS

For normal 512x512 tiles the old ones load in 13-15s for me and the new ones in 3-5s. Analytical download 1250x933px of the same location (zoom=12&lat=43.02109&lng=11.73855) took 20s vs 1.1min.